Restructured aligned by rng as master.
