+++ .travis.yml
@@ -4,12 +4,14 @@ php:
   - 5.6
   - 7.0
   - 7.1
+  - nightly
   - hhvm
 
 sudo: true
 
 matrix:
   allow_failures:
+    - php: nightly
     - php: hhvm
   fast_finish: true
 
@@ -30,7 +32,7 @@ script:
 
 after_script:
   # We upload only for reference platform! This is our base for further code analyses and so on.
-  - if [ $(phpenv version-name) == "5.6" ] && [ "$PREFER_LOWEST" == "--prefer-lowest" ]; then bin/codacycoverage clover build/logs/clover.xml -vvv; fi
+  - if [ $(phpenv version-name) == "5.6" ] && [ "$PREFER_LOWEST" == "--prefer-lowest" ]; then bin/codacycoverage clover build/logs/clover.xml > /dev/null 2>&1; fi
 
 after_success:
   # Push coverage to github pages branch
@@ -43,7 +45,7 @@ after_success:
   - echo -n $GIT_TAG > public/version
   - git commit -m "Set build VERSION number" public/version
   - git tag $GIT_TAG -a -m "Generated tag from TravisCI build $TRAVIS_BUILD_NUMBER"
-  - git push --tags --quiet https://$GITHUBKEY@github.com/clickalicious/webserverdaemon > /dev/null 2>&1
+  - git push --tags --quiet https://$GITHUBKEY@github.com/clickalicious/webserver-daemon > /dev/null 2>&1
 
 # blacklist
 branches:
@@ -53,7 +55,7 @@ branches:
 # Who to notify?
 notifications:
   slack:
-    secure: jg72gIt0LLN0bc8QnILaYuSDpxnrsPjvX1nXzZqY9z6YFEBLPNmPZgPf3R1EDMesJDbUEAWD/s60bhBa2B9UjxT1Cr6wxHuisuiAEMjbQn9CkQHy+LDp9lyj+KzY+3ZmelLsQXbU3v9mSF1aGQqmW46GGR21NDDBkF2+9u28izOV0qE/086jKW97OQx9099OTl85BkQgV69lIG/Lq09kCL0C6ErIPe2IBsFuFDMLDHAJJzwOvfvJ2QB1snjriKI5acu76dOeGr5dZF4CTFHxnh2AZj2earzUIVTDB7aADaQa7/Goc38wim3r2kEt00Nni70wT8Ztk865fIfXFejgykYiDnlg+oIpfEyZwqpYWxX+O3hJmAz9skick/0UZ4HuvxLM94N9aex2LuSONOutf2UmaRfPj6VUqjFisBDOuVR2qrNAKuvZxVy30szLvyGNmqUvLIvZVVO/qMg1reEBAqyXK0UH3w0RsKpLzj92NtEAab4NtLUQZ6Ne7+ouJQ7v8y2/QdQg6Mf2Qth4sBgTXmNGqiZ0UJGtWJumfTH0sIneqnM5qkYl7GIkpy/PBUU5Cv1mDqk+Q4dd27vs9A7ptDXnnPSv1Nx4Z+6v5Ut8ZwgxDpgvZMNl1nQwAJD6v/i1TCNl6buny7dx1HvvX5v/GPY/8m2R+9soDg5pkWzsEwQ=
+    secure: "jg72gIt0LLN0bc8QnILaYuSDpxnrsPjvX1nXzZqY9z6YFEBLPNmPZgPf3R1EDMesJDbUEAWD/s60bhBa2B9UjxT1Cr6wxHuisuiAEMjbQn9CkQHy+LDp9lyj+KzY+3ZmelLsQXbU3v9mSF1aGQqmW46GGR21NDDBkF2+9u28izOV0qE/086jKW97OQx9099OTl85BkQgV69lIG/Lq09kCL0C6ErIPe2IBsFuFDMLDHAJJzwOvfvJ2QB1snjriKI5acu76dOeGr5dZF4CTFHxnh2AZj2earzUIVTDB7aADaQa7/Goc38wim3r2kEt00Nni70wT8Ztk865fIfXFejgykYiDnlg+oIpfEyZwqpYWxX+O3hJmAz9skick/0UZ4HuvxLM94N9aex2LuSONOutf2UmaRfPj6VUqjFisBDOuVR2qrNAKuvZxVy30szLvyGNmqUvLIvZVVO/qMg1reEBAqyXK0UH3w0RsKpLzj92NtEAab4NtLUQZ6Ne7+ouJQ7v8y2/QdQg6Mf2Qth4sBgTXmNGqiZ0UJGtWJumfTH0sIneqnM5qkYl7GIkpy/PBUU5Cv1mDqk+Q4dd27vs9A7ptDXnnPSv1Nx4Z+6v5Ut8ZwgxDpgvZMNl1nQwAJD6v/i1TCNl6buny7dx1HvvX5v/GPY/8m2R+9soDg5pkWzsEwQ="
   email:
     recipients:
       - opensource@clickalicious.de
