+++ LICENSE
@@ -1,7 +1,5 @@
-webserverdaemon
-
 (The MIT license)
-Copyright 2017 clickalicious UG, Benjamin Carl
+Copyright 2017 clickalicious, Benjamin Carl
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the
